+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__/
 .pytest_cache
 .vscode
 env.sh
+venv/
